Lint check for explicit hiera() usage
As explained in the comments. The plugin doesn't add any appreciable
difference to the time that `bundle exec rake lint` takes.

A minor change will be required when puppet-lint 0.4 is released because
checks are defined in a different way (`PuppetLint.new_check`).

We should work on removing the exemptions listed under the FIXME.