[#66656002] Use underscore class names for logship
Change logstream/logship to use an underscore instead of a hyphen in the
class name when calling `govuk_node_list` to find the current `logs-redis-N`
machines. Because hyphen support was removed in c01fcf6.

This resulted in recently reboot machines, like `backend-1.backend.preview`,
not being able to send logs because the string evaluated to empty. Hence the
`-s redis,,` argument seen here:
```
 3608 ?        S      0:00 /usr/bin/python /usr/local/bin/logship -f init_txt,add_timestamp,add_source_host,add_tags:stdout:upstart,add_fields:application=transition -s redis,,key=logs,bulk=true,bulk_index=logs-current
```