[#66672712] Mongodb should start after networking
This is a totally speculative fix for #66672712 which I am unable to
reproduce locally. We have seen several times that mongodb on real
machiens fails to start after a reboot.

About the only case I can think of would be if it's waiting on a
filesystem or networking in order to start correctly and without it, it
shuts down.

By making it start after networking, we guarantee that it is starts
after the network (and filesystem, which is an earlier step) is up.

The start line was cribbed from `/etc/init/mountall-net.conf` and the
stop line was cribbed from `/etc/init/ssh.conf`. Both lines were
verified to work on both Ubuntu 10.04 and 12.04

However regardless of this fix, I'm going to close off [the
story](https://www.pivotaltracker.com/story/show/66672712) as unable to
reproduce, because I can't.