Convert collectd Redis plugins to classes.
As it stands, they would break if instantiated multiple times, so they
don't make sense as defined types.

At some point, we may well want to be able to specify extra queues at
different points in the hierarchy, rather than having to specify them
all in one place. When that happens, we can factor out the queue plugin
into something with a defined type in it.