Add external-link-tracker app
This is a Go application that has similar configuration to the Router.

It has a public component intended to be accessed via GOV.UK, which
proxies to the backend VDC. It needs to reside in the backend because it
writes to a database.

This pattern is based on the publicapi module we use to proxy the
content API publicly.

We override the default Nagios healthcheck because we want to listen on
the API port - this is the part of the application that serves the
`/healthcheck` path.