[#66760818] Don't check lost+found.
`CHECK_LOSTFOUND="no"` in `/etc/default/cron` will cause `/etc/cron.daily/standard`
to not check for the existence of `lost+found` in mounts.
It will also make `/etc/cron.daily/standard` to not check for any files in
`lost+found` if it does exist.

We have decided we don't care mainly because if it does exist in /var/lib/mysql
mysql will think its a database.