[#66664898] Remove remaining check_redis alerts
We no longer use this plugin for checking list lengths (c01343e) or memory
usage (af91ab4). The remaining checks are:

- Connection latency / response time

  Looking at the code it appears to be the elapsed time in which the `INFO`
  command returns. Rather than the avg/max response time for other clients
  talking to Redis. Which I can't see has ever fired an alert for us and
  isn't incredibly useful.

- Connection clients

  The limit for Redis 2.4 (which we're using) is hard-coded to 10k clients.

  - http://redis.io/topics/clients
  - https://github.com/antirez/redis/blob/2.4.13/src/ae.h#L36

  Our stats in Graphite show that we hover at a couple of hundred. Again, we
  never appear to have fired this, so I don't think it's terribly useful.

I think these were all added because we just wanted to monitor *something*
when we first introduced Redis to our stack. But they aren't providing us
much value and the plugin is incredibly difficult to use/read. So remove
them.

The `icinga::nrpe_config` and `icinga::check` resources should purge
themselves. `icinga::plugin` won't because there are unmanaged plugins in
that directory, so it's marked as `absent`. The `cpanm` provider doesn't
allow us to remove the `Redis` package, but that's probably not a big deal.