alert for 501s from bouncer
A 501 from bouncer generally indicates a bouncer misconfiguration -- an
invalid redirect, a missing domain from a whitelist, or something like
that.  While this Shouldn't Happen, it occasionally does, and it would
be good to feed back to @jamiecobbett and co that this is happening.

Used a 1 hour time window so that the alert will stay visible for a
while rather than blinking on and off.  Used a low threshold for warning
and critical because we shouldn't get *any* 501s at all, ever.

TODO: also write opsmanual entry for what to do when this alert
triggers. Which will probably be "tell the transition team".