[#66937478] fix check_elasticsearch_syslog_input threshold
This fixes check_elasticsearch_syslog_input to fail properly when it
should, and not to when it shouldn't.

I think I've worked out why 243af9cd29ff1 ever happened: it seems that
@0 simply does not work as documented: it *always* fails:

    ppotter@preview-monitoring:~$ /usr/bin/check_graphite -H http://graphite.cluster/render -M 'transformNull(removeAboveValue(carbon.agents.graphite-1-a.errors,-1),0)' -w @0 -c @0 -F 60sec
    CRITICAL: value=0.0|value=0.0;;;;
    ppotter@preview-monitoring:~$ /usr/bin/check_graphite -H http://graphite.cluster/render -M 'transformNull(removeAboveValue(carbon.agents.graphite-1-a.errors,-1),1)' -w @0 -c @0 -F 60sec
    CRITICAL: value=1.0|value=1.0;;;;

What we want is to fail outside the range (0,inf) (excluding both
endpoints), which is naturally expressed as 0: (which @tombooth tried to
do) but unfortunately this *includes* both endpoints so it *never*
fails:

    ppotter@preview-monitoring:~$ /usr/bin/check_graphite -H http://graphite.cluster/render -M 'transformNull(removeAboveValue(carbon.agents.graphite-1-a.errors,-1),0)' -w 0: -c 0: -F 60sec
    OK: value=0.0|value=0.0;;;;
    ppotter@preview-monitoring:~$ /usr/bin/check_graphite -H http://graphite.cluster/render -M 'transformNull(removeAboveValue(carbon.agents.graphite-1-a.errors,-1),1)' -w 0: -c 0: -F 60sec
    OK: value=1.0|value=1.0;;;;

The hack here is to exclude zero by using a very small nonzero number,
expressing the range as '0.000001:'. This fails when zero, and passes
when nonzero:

    ppotter@preview-monitoring:~$ /usr/bin/check_graphite -H http://graphite.cluster/render -M 'transformNull(removeAboveValue(carbon.agents.graphite-1-a.errors,-1),0)' -w 0.000001: -c 0.000001: -F 60sec
    CRITICAL: value=0.0|value=0.0;;;;
    ppotter@preview-monitoring:~$ /usr/bin/check_graphite -H http://graphite.cluster/render -M 'transformNull(removeAboveValue(carbon.agents.graphite-1-a.errors,-1),1)' -w 0.000001: -c 0.000001: -F 60sec
    OK: value=1.0|value=1.0;;;;

Bah humbug.

(and yes, those test lines to check_graphite are stupid. I should be
able to just directly test against constantLine(0) or constantLine(1)
but that causes a ZeroDivisionError. So I take an existing metric,
remove all its values, then transform null to my target test value.)