Add ensure to govuk::app
Add ensure to govuk::app, govuk::app::config, gouvk::app::envvar,
govuk::app::nginx_vhost and gouvk::app::service.

This change mostly just ensures that files are absent. It also ensures
that the service is stopped. The base service provider is used so that
we don't have to make sure the service is stopped before the upstart
config file is removed.