[#67094978] Prevent random Hiera test failures
Puppet's spec tests were failing intermittently due to an apparent race
condition whereby the run order of these tests would sometimes trigger a
failure.

By defining a Facter fact that changes for each test, we force
'rspec-puppet' to invalidate cached Puppet catalog and the tests pass.

This can be removed once this issue is resolved in
'hiera-puppet-helper', which we're using for `:hiera_config`.