validate govuk::logstream ensure parameter, restore ukba logstream
The govuk::logstream changes introduced in 7d625cb change the `ensure`
parameter to expect `present` or `absent` as parameters.  Due to the
ordering of PRs, this didn't catch the previous usage of `true`
introduced in b01b6af3.  This meant that the ukba vhost logstream was
removed, because the logic meant that `true` was interpreted as
`absent`.

This introduces some proper validation and fixes the ukba logstream to
reappear.

cc @jennyduckett