Fix problem with duplicate "fields" entries in kibana2_url
The default was specified as the symbol :fields, while the parameter
passed contained the string 'fields'. Both can exist in the same hash,
which when JSONified results in a hash containing both keys. (The test
data even required the duplicates to be present...)

Since puppet doesn't know about symbols, the fix is to just use strings
everywhere.