Add "offset":0 to default kibana query string
If you don't specify an offset value, you end up with a search range of
"undefined to NaN", and also the pagination links to older log lines
disappear. For example:

https://kibana.production.alphagov.co.uk/#eyJmaWVsZHMiOltdLCJzZWFyY2giOiJAZmllbGRzLnN5c2xvZ19wcm9ncmFtOlwicHVwcGV0LWFnZW50XCIgQU5EIEBzb3VyY2VfaG9zdDpcImNhY2hlLTFcIiIsInRpbWVmcmFtZSI6IjE0NDAwIiwiZ3JhcGhtb2RlIjoiY291bnQiLCJ0aW1lIjp7InVzZXJfaW50ZXJ2YWwiOjB9LCJzdGFtcCI6MTM5NDE5OTYwMTcwMX0%3D

This adds "offset":0 to the default query string to stop this problem.

Also, the emacs functions base64-{de,en}code-region are awesome for
editing these tests :D

http://www.gnu.org/software/emacs/manual/html_node/elisp/Base-64.html