collect ukba error log and statsd metrics from json event log
Working to follow the pattern in nginx::config::vhost::proxy.

statsd_metric and statsd_timer set in order to collect graphite metrics
around status codes and response times.  This will automatically be
picked up by the bouncer 501s check introduced in 04fc281, due to the
format of the wildcard.

Also collect the error log into kibana.  The error log contains
information that the access log doesn't contain, so it's important we
collect both.