Toggle using NFS for synced folders
Applies the same behaviour as this commit in the 'vagrant-govuk' repo:

https://github.com/alphagov/vagrant-govuk/commit/c818e8f6bfdd7c98ce7ccf3142420a6a9166b6b1

While NFS is more performant it has issues when using at least one of
our VPNs. The performance gain is not as useful if you are only checking
puppet runs.

Add an environment variable VAGRANT_GOVUK_NFS to toggle the use of NFS
for synced folders.
Default to using NFS.

If you dont want to use NFS set VAGRANT_GOVUK_NFS=no as an environment
variable.

e.g

$ VAGRANT_GOVUK_NFS=no vagrant up box.example