Remove tariff_production database from puppet
Trade Tariff was changed to a temporal database model (whatever that means)
and during that change was migrated to the
`tariff_temporal_production` database.

As per alphagov-deployment:
```
[gds/alphagov-deployment:master●]$ ack tariff_production trade*
[gds/alphagov-deployment:master●]$ ack tariff_temporal_production trade*
trade-tariff-backend/to_upload/database.sky.yml 3:  database:
tariff_temporal_production
```

This commit removes the database from all the machines. It leaves a
couple of FIXME's to be done at some point later.

This is related to Pivotal [#67106956].