Remove mongo lock percentage check
This is not a helpful check.  It divides total lock time since startup
by total time since startup.  Kill it with fire.

There is an alternative: we could use the current_lock check from
check_mongo.py.  That uses a state file in /tmp to keep track of the
previous total time and lock time, finds the delta, and from this
determines the lock percentage since the state file was last updated.

I don't think we care enough about this to warrant using it.