Remove old router fallback page
This functionality comes from "Initial commit" and allows a static HTML page
on the cache nodes to be served for all GOV.UK requests. It predates having
an external mirror and I'm not aware of being used for as long as I've been
here. Nowadays we would just disable the vhost on the cache node and let the
CDN send requests to the mirror.

Removing this makes the vhost slightly simpler to read. It will also reduce
unnecessary `stat(2)` calls for every(?) request.