Sync asset-manager directories to asset-slave
This is a bit of a whoopsie.

There was an out of data copy of the asset-manager directory on the
Production asset-slave. When syncing data, this gets copied to the
asset-master in the destination environment. There's just enough in it
to pass the asset-manager smokey checks.

However if you do a sync from that destination environment to a new
environment (e.g. P1 Production to P1 Staging), nothing is transferred
and the asset-manager smokey checks will fail, because it isn't copied
from the master to the slave on the first destination, so isn't
available when syncing from slave in the first destination to master in
the second destination.

This simple rsync job will make sure that the assets are the same on the
master and the slave in the environment.

/cc @alextomlins