[#67234604] Set a global TZ environment variable
Setting this environment variable will prevent Ruby applications from
unnecessarily calling stat(2) on `/etc/localtime` on *every* `Time`
operation.

Ref: http://www.chemie.fu-berlin.de/chemnet/use/info/libc/libc_17.html#SEC301

Arguably this shouldn't make much difference to our IO usage because the
kernel should cache small and frequently used files like this. But it will
make the output of `strace(1)` more usable.

This won't restart any applications to pick up the change. It might be
useful to restart everything in Preview, where uptime isn't a big deal and
we can observe any differences. But we can just leave Staging and Production
to pick up the change as apps are deployed.

The value is set to UTC because that is the One True Timezone (and it
matches the `timezone` module which managements `localtime`).