Flag destinations of whitehall-mysql aliases
In platform-one Whitehall uses different MySQL servers. In order to
homogenise the branches of alphagov-deployment, we need an alias for
Whitehall's MySQL servers which points to the respective servers, such
that we can have a consistent database.yml that will work for Preview
(remaining on Interim Platform) and Production/Staging depending on
whether we have migrated yet.

This feature-flags the legacy_aliases of the respective hosts such that
on the Interim platform, an additional alias is added to the shared
mysql servers which Whitehall will point to. On Platform 1 this will
point the dedicated MySQL cluster.

The default for `hosts::production::whitehall_shares_mysql` is currently
defaulted to `true` to match the behaviour of the Interim Platform. In
the `deployment` repo the correct behaviour will be set per-platform.