[#67649440] Manage (and upgrade) Hiera at v1.3.1
By not previously managing this we have a mixture of machines with different
versions depending upon when they were created:

- puppetmaster-1.management.preview: 1.3.0-1puppetlabs1
- puppetmaster-1.management.staging: 1.2.1-1puppetlabs1
- puppetmaster-1.management.production: 1.2.1-1puppetlabs1
- puppetmaster-1.management.p1staging: 1.3.1-1puppetlabs1
- puppetmaster-1.management.p1production: 1.2.1-1puppetlabs1

Upgrade them all to 1.3.1 since it appears to be working fine in Platform1
Staging and Vagrant. This will restart the PuppetMaster service.

Also pin our `Gemfile` to match (although it wasn't far out).