[#67649440] Monkey patch for Puppet 3.4 upgrade
During testing I've seen the following errors intermittently:
```
Error: Could not back up /usr/bin/puppet: uninitialized constant Puppet::FileSystem
```
```
Error: Report processor failed: uninitialized constant Puppet::FileSystem
```

As described in PUP-1603 it seems that when Puppet upgrades itself to 3.4.x
some parts of the new codebase are picked up but others aren't. Thus it
doesn't know about the new module `Puppet::FileSystem`. Although I can't
reproduce this reliably, the worst case scenario is that it leaves Puppet in
a broken state (such as not restoring our custom binary).

Workaround this by providing our own `Puppet::FileSystem::File` for
pre-upgraded clients which will act as a shim for the first call. It will
log a message, load the correct version from 3.4 (replacing our class for
future calls), and proxy the one call that we caught to the real class. We
also have to patch `Puppet::FileBucket::File` because the new
`FileSystem::File` refers to two new methods in it.

This isn't *so* evil because it will no longer be active once a machine has
been upgraded to 3.4 and we can remove it altogether when we stop
provisioning/bootstrapping new machines with older versions of Puppet.