[#54693536] Remove old puppet gems from bootstrap
Machines in Skyscape Interim and Vagrant have Puppet, Facter and Hiera
installed as gems into the system Ruby by gds/vcloud-provisioner and
alphagov/boxes (respectively). But not Skyscape Platform1 because
gds/govuk-provisioning uses the Debian packages from the beginning.

This hasn't caused too much of an issue since we upgraded to Puppet 3.2.x
because we also bumped the gems (mostly by hand) to the same version.

However they are otherwise unmanaged and when we come to upgrade to 3.4.x
we'll see the following error on the Puppet master because it's using the
wrong codebase:
```
Error: Could not retrieve catalog from remote server: Error 400 on SERVER: Could not intern from pson: source '%7B%22name%22%3A%22p' not in PSON!
Warning: Not using cache on failed catalog
Error: Could not retrieve catalog; skipping run
```

So remove them. We can't use a package provider because their names conflict
with other packages and we need to retain the executables that are now owned
by the debs. Removing them in reverse dependency order. It must happen after
`puppet-common` or Puppet will complain that is has no code underneath it
anymore.