[#67404250] Reduce Graphite disk IO
Reduce the frequency at which Graphite's carbon daemon writes to whisper
files on disk. This should help us (and other tenants) on IO-challenged
vendors.

Carbon buffers the additional data points in memory and calls
`update_many()` less frequently but with more data each time. This is more
efficient because it results in less disk operations and less time jumping
around the disk for random writes.

I have been testing this on Preview where we have 155k metrics. A value of
500 had little or no effect. Dropping it to 200 made a big difference, as
demonstrated by:

- Disk writes down from 1500 IOPS to 750 IOPS.
- `pointsPerUpdate` has increased from 2.5 to 13. Because each
  `update_many()` now contains more data.
- `updateOperations` has dropped from 60k to 12k, as a consequence
- `committedPoints` now fluctuates more around the value of
  `metricsReceived` which is expected as the buffer fills and flushes.

I'll throw a copy of these graphs into Pivotal for future reference. This
value seems like a good compromise, so I'm not going to experiment any
lower/higher.

I'm uncommenting `MAX_UPDATES_PER_SECOND_ON_SHUTDOWN` and leaving it at the
value that `MAX_UPDATES_PER_SECOND` used to be so that we can quickly burst
any buffered metrics to disk when stopping or restarting carbon.

All credit to @rolandpienaar for implementing this on IDA and telling us
about the results.