Configure contacts app for frontend and backend
This follows the Whitehall pattern of having a single app which is used
for both admin and frontend talking to the same MySQL database. This
pattern is not desirable.

After getting the app launched to unblock HMRC transition we plan to add
a contacts API and make the frontend use that instead of using MySQL.