Swap ordering of puppet/puppet-common packages
The `puppet` package places an APT dependency on a specific version of
`puppet-common`. When the `$puppet_version` variable changes and Puppet
calls `dpkg` to install a new version of `puppet-common`, the `puppet`
package gets uninstalled and so does `puppetdb`. It seems that `puppet`
sometimes gets reinstalled in the same run but PuppetDB is left broken until
the next run - possibly because the list of installed packages is prefetched
and not re-evaluated.

By reversing the order we can ensure that they get upgraded correctly in a
single run and without uninstalling any dependencies. I didn't mention
anything in my original commit message for these packages, so I'll assume
that there was no good reason for their order being that way previously.

The resources within this class which depend on `puppet-common` should
continue to do so, since that signals a complete upgrade/install.