Fix dependencies on puppetmaster cert
Don't [re]start Nginx or run the PuppetDB command which copies the SSL cert
until we have created one. This allows us to bring up a new puppetmaster in
a single run.

Unfortunately there's no nice way to indirectly notify Nginx if the
certificate changes but the vhost doesn't. But since we've never had this
situation so far; I'm not going to worry about it for now.