Run smokey with the correct environment
Smokey on the Monitoring machine relies on a number of environment
variables to affect it's operation. If they aren't set, it defaults to
checking Preview - which is wrong.

In particular GOVUK_APP_DOMAIN and GOVUK_WEBSITE_ROOT are not set in the
environment. When running post-deploy smoke-tests from Jenkins, the
commands are prefixed with `govuk_setenv default` which sets the
variables correctly.

This commit makes the upstart job which runs Smokey on a cycle also use
govuk_setenv, which should mean the checks run against the correct
platform.