Temporarily lower the asset space alert threshold
At the moment we have a some acknowledged alerts for space on
/mnt/uploads on the asset machines in Production. These will be resolved
when we move to Platform 1. In the meantime, we won't realise if
something increases our space usage, which is dangerous.

By lowering the thresholds on this alert, it will resolve now and then
trip again if the space usage falls below the new limit.

This commit can be reverted after Platform 1 migration and I will create
a story to make sure that happens.