[#67404250] Patch for Graphite->Carbon comms
The upgrade to Graphite 0.9.12 broke one of the ways that the Graphite
webapp obtains metric data for display. It was still reading the vast
majority of metrics from Whisper files on disk. But it stopped reading the
most recent data, which hadn't been persisted to disk yet, from
carbon-cache's buffer.

This wasn't apparent to us until we reduced the amount of disk I/O that
Carbon was consuming in 49a9861. When we stopped writing to disk so
aggresively, and the Whisper files were less up-to-date, we noticed that the
graphs were incomplete for the most recent ~3 minutes. On closer
inspection there were also lots of errors in `webapp/exception.log`.

Besides making our graphs look Not Right, this will have a negative effect
on our Icinga `check_graphite` alerts, because they no longer have access to
accurate data for the last five minutes (the default check timeframe).

This has been fixed in the Graphite 0.9.x branch but not yet released (as
what will be 0.9.13). I'm backporting the fix because of how seriously it
affects us. There's no nice way to do this, but this should at least:

- Be obvious about what patches we've applied from upstream
- Blow up if we change the version but don't remove the fix

I've pinned the Graphite version to make this more obvious.