Remove Nagios alerts for Fastly error rate
While these are something that we certainly should care about, in practice
the alerts are of limited use because Fastly only provide stats for 5~15
minutes behind the present time.

This means that the alerts have been firing after we've already been made
aware of an outage by other means (Pingdom/Nimsoft, 5xx within our stack,
etc) and they usually persist for a short while after we've resolved the
outage. We're also unlikely to get any better data from other CDN providers.

We now have this data in Graphite - with the correct (-5~15mins) timestamps.
So we can still persue it at our own leisure for historical info.

As discussed with @samsharpe and @mattbostock