Reload puppetmaster on hiera.yaml changes
This doesn't happen by default because of the afore-referenced bug. This
makes deploying changes to the hierarchy very troublesome.

Workaround this by having the agent track the checksum of the file on disk,
but no other properties. We have to `undef` a bunch of them because we set
defaults in `site.pp`. When the checksum changes the `::config` class will
notify `::service` to trigger a reload.

This still isn't perfect as described in the comments. But I think it's
better than the previous situation. The only alternative to this is to write
some logic in our Capistrano job that restarts the service if deploying
changes to certain files. However the version of `railsless-deploy` that we
currently use doesn't support `deploy:pending:diff`, it wouldn't work in
Vagrant, and I think it might be more confusing to have the logic in another
repo.