Reload puppetmaster on function/type changes
Same as fbb1e4a but for Puppet functions and types that are distributed from
modules using `pluginsync`. We don't care about other sorts of plugins like
facts and providers because they are loaded by the agent rather than the
master.

This results in approx 100 more resources on the catalog at the moment.
However the checksumming appears to be very quick and adds no appreciable
time to the Puppet run.