[#68258934] Global known_hosts for offsite-backup
Add a global `known_hosts` entry for the current offsite-backup machine in
Memset. Populate the current host key in the common hieradata (it's not
sensitive and doesn't change between environments).

This will allow backups to work on Platform1 Production where the host key
hasn't been manually accepted. It should also allow us to change the
host/key if necessary in the future.

The ssh client will still add the IP (hashed) to `~/.ssh/known_hosts`. This
is in part why I'm putting it into the global hosts rather than managing
that file. I'm not 100% sure what will happen if we change the IP. It will
either break and we have to accept the new IP, or it won't break and we
don't have to do anything.

I'm also copying some dependency related stuff from the `backup::server`
test since `backup::offsite` also relies on the user from `backup::client`
but it isn't immediately obvious from the manifest.