Add more directories that Whitehall requires
We precreate some directories that Whitehall needs - these appear on
both the asset slave and the asset master. Whitehall will create other
directories in deploy:setup and initialization, where we don't
necessarily need the contents to be synced between Master and Slave,
so they are not synced from master to slave via cron.

This leads to the problem that when we sync from the Asset Slave in one
environment to the Asset Master in another environment (with the
--delete) option, we remove some directories that are actually needed
by the running application. Oops.

This makes sure the necessary directories are created by puppet on both
the asset master and slave, which will lead to a successful sync
between slave to master.

It also [Delivers #68096110] which I actually manually fixed on the P1
Production slave which isn't a sustainable solution.