[#68276766] Change gor to Platform1 staging
Now that we've migrated Production and Staging to Skyscape Platform1 the old
targets of Interim Staging and Platform1 Production no longer make sense.

Repoint "staging" to Platform1 Staging, so that we can enable traffic replay
again. Remove the target and feature flag for "plat1prod", since this is no
longer a thing (we don't want Prod to replay against itself).

Make changes to host entries notify the gor class which will restart the
service. We don't need this right now, because `enable_staging` isn't
enabled anywhere. But it may be useful in the future.

The old `host{}` entry for "p1prod" will get purged by 6da6025.

I've left some of the logic for specifying multiple hosts in case we want to
do something similar again in the future. I've had to remove the spec test
though, since this no longer applies.