Amend monitoring hostnames
Since we no longer use `alphagov.co.uk` subdomains as part of the naming
convention for FQDNs attached to supporting services (yay!), and since we now
have a new Puppetised back-up machine, this commit amends the hostname to
`backup0.backup.production.govuk.service.gov.uk`.

A new naming convention is being applied to the `$offsite_hostname` variable; we
do not currently have a convention for the shorthand notation of infrastructure
tools on `service.gov.uk` domains. This commit defines the convention, which
will be to take the first and third segments of the full URL.