[#54520168] Use mirror of apt.puppetlabs.com
Use our internal mirror for Puppet packages. Depends on gds/deployment#201

IMPORTANT: The firewall restrictions on `apt.production.alphagov` mean that
you will no longer be able to provision Vagrant boxes when not in the office
or on the `/ah` VPN profile.

Includes a feature flag that will flip this back to the upstream repo. Which
we might need to use if we were bringing up a new environment that replaced
or couldn't access the current Production.