[#68696262] aptly::mirror for GOV.UK PPA
We need one entry for each distribution.

Aptly has a special `ppa:user/project` syntax that you can use here which
will use the config options `ppaDistributorID` and `ppaCodename`. But for
simplicity, and because we need two distributions, I'm just using the URL
that appears in the APT sources list.