Amend offsite-backup script template
This commit amends the offsite backup script such that it now:

- creates $TEMP_FILE using mktemp
- writes the GPG-secured, encrypted tarball to $TEMP_FILE in /data/backups
- SCPs $TEMP_FILE to the remote machine
- removes $TEMP_FILE from the local machine
- make the @dest_folder a class parameter

Since the rest of the block in which the back-ups are executed is parameterised,
either using variables where the value is passed locally, or using class
parameters where the value is passed through Hiera (but where they are presented
as Ruby variables), we should also make the folder on the destination service a
class parameter.

A separate pull request in alphagov/govuk_offsitebackups-puppet will allow SCP
access through RSSH for the govuk-backup user.