Provide a default value for $dest_folder
If we do not provide a default value for the $dest_folder class parameter, the
existing off-site backup script will break unless we provide a rather
blank-looking value in Hiera, akin to `backup::offsite::dest_folder: ''`, which
is messy.

Instead, we make a default value here, which can then be overridden by the Hiera
key/value pair when the new off-site backup machines are deployed. This key will
be included in #1618.