Secure the nginx ssl keys with mode 0640
This partially [Fixes #67015586] which notes that our NginX SSL keys are
world readable.

The story mentions setting the directory so that it is not world-readable
too, but I have taken the decision not to do that because it also contains
the certificates and it would be possible to derive the list of keys and
certs from the NginX config anyway, so I don't think it's a security
risk.

PS. I have extended the test to catch this regressing. Woohoo!