Add dest_folder key to Hiera
This commit adds the dest_folder key to Hiera. By including it here, we can
ensure that this value only starts to be used when this pull request is merged
and deployed. Using it earlier than this results in a sloppy Hiera dataset, and
might break the existing off-site backup script.