Re-work SSH_OPTS for use with SCP
This commit re-works the SSH_OPTS variable for use with SCP. The options
provided are:

C, for compression, which is always a Good Thing (TM).
p, to preserve modification times etc., which is good to keep.
o, to provide custom SSH options

VisualHostKey=no, stops the visual hostkey appearing, which quietens cron.
Quieting cron is a good idea because else it devalues the mail which appears in
machine.email@d.c-o.g.u. Cron will still mail the list when an error occurs
though, which is good.