[#63789506] Use new mirror of S3 repo
This is now split into two repos. Lucid backports should only be available
to Lucid machines. We can remove the `apt::pin` resources we were
previously using to enforce this.

I'm not providing a feature flag to disable this repo. If it goes down then
we should just disable it, rather than reverting to the original S3 bucket
which may no longer exist in the future.