Upgrade to Facter 1.7.5
It seems that a problem with mirroring the "stable" distribution of the
PuppetLabs repo in d2ab45dae9efd58be240138ec0f5c05a7e077a6a is that they
evict old packages more frequently than the "precise" or "lucid" dist paths.

This version of Facter doesn't exist in stable. So bump to the latest minor
release, which is, and we've mirrored.

- http://docs.puppetlabs.com/facter/1.7/release_notes.html#facter-175

Snapshots should prevent this from happening too often. It still feels like
"stable" is the right thing to use, rather than "precise" everywhere.