Allow govuk::app to depend on NFS mounts
Some applications depend on mounted NFS shares for full functionality.
If these applications start before the NFS share has mounted (as
sometimes happens during a reboot), then we see an increase in error
rates.

This change allows a govuk::app to depend on whether the upstart task to
mountall-net has run. This task is triggered on net-device-up, so in
theory it will only start this job once net-device-up has triggered and
then mountall-net has triggered. mountall-net runs and exits, which is
why starting on 'stopped mountall-net' is the right thing to do.

I'd appreciate a review of my upstart logic and the work on govuk::app which I
haven't looked at for a very long time. The upstart cookbook entry that
I am working from is: http://upstart.ubuntu.com/cookbook/#event

This is for [#55487134] but it will not be fixed until the logic is
tested and the necessary applications are configured in puppet to use
it. The applications we know about are:

- asset-manager,
- whitehall on backend servers
- possibly contentapi)