Disable atime/relatime in Prod/Staging Graphite
This commit implements the same as  6be347a but for Production and
Staging. Implements [#68966642].

* * *

Disable atime[1] and relatime[2] for Graphite storage to help reduce the
number of IOPs.

As of Linux 2.6.30, all filesystems are mounted with the `relatime`
option[3] which means that the file access time is only updated when a
file is written to. Setting `relatime` still triggers a write to disk
each time a file has been modified[2], which is undesirable for the
Whisper files used by Graphite.

The change here will disable both atime and relatime.

[1]:
http://tldp.org/LDP/solrhe/Securing-Optimizing-Linux-RH-Edition-v1.3/chap6sec73.html
[2]:
https://access.redhat.com/site/documentation/en-US/Red_Hat_Enterprise_Linux/6/html/Power_Management_Guide/Relatime.html
[3]:
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=0a1c01c9477602ee8b44548a9405b2c1d587b5a2