[#68359120] Don't use the logical::volume defined type.
The logical::volume defined type is a handy shortcut for the case of a
volume group having one physical volume and one logical volume.
This is less than helpful if you want to have more than one of either
associated with a volume group.

This commit makes the govuk::lvm defined type use the underlying
providers to give us more flexibility.