[#62542906] Alert thresholds on origin 5XX errors
Reduce the threshold at which Icinga will alert on HTTP 5XX errors at
origin, AKA the cache nodes.

These values are based on the per-application thresholds, currently set
to 0.05 for warnings and 0.1 for critical alerts. The threshold here is
more tolerant as we could see multiple applications spike
simultaneously during normal operations.

Note that under the current configuration, either a warning or a
critical alert will cause second line support to be paged.