[#68359120] Tests for multiple physical volumes
The refactoring of this class to not use lvm::volume now means that we can pass
it an array of physical volumes to make up a volume group.

These tests check that the volume group class exists with the array of physical
volumes and that all physical volume classes exist.

Adding these tests because @bobwalker and @calpaterson arn't entirely sure that
what we're doing is expected behaviour of puppet.