Whitehall depends on NFS
Whitehall on the backend servers depends on NFS so this gives it a
chance to wait until NFS has had a chance to mount before it tries to
start. Note that whitehall on the frontends is not technically dependent
on NFS, but slowing down it's startup fractionally in the unlikely event
that all the servers reboot at the same time is not a bad thing.

It in part [Fixes #55487134] and is dependent on #1633