[#69443868] Only download key if key != pre-exist
After the key has been downloaded successfully for once, and therefore appears
in the list of keys which GPG returns, ensure we don't re-download they key as
this requires us to depend on key servers being available.

On 12th April 2014, a backup did not run successfully because the keyserver was
not contactable.