[#66933488] Fix Puppet passive check freshness
Add a `set -e` to abort the script if `puppet agent` returns a non-zero exit
code. This prevents it running `puppet_passive_check_update` against an old
`last_run_summary` file, submitting a bogus "OK" result to Nagios, and
causing it to never exceed the `freshness_threshold`.

This recently happened on `asset-slave-1.backend` in the referenced story
because an old `agent_catalog_run.lock` was left and prevented any
subsequent Puppet runs. Nagios thought everything was fine.

Per the inline comments, we need to mask an exit code of 2 because it's
possible to pass `-v` as an additional argument to `govuk_puppet` which
enables `--detailed-exitcodes`:

http://docs.puppetlabs.com/references/3.stable/man/apply.html#OPTIONS

I've also snuck in a `set -u` for safety against unset variables. Since
there are a few vars used in this script.