[#63211736 #63211650] Cache all librarian modules
So that we don't download other peoples code from the internetz on every
deployment. In other words protecting us from:

- Other people being able to change the code we run (as root)
- Us being unable to deploy due to Forge or GitHub downtime
- Sloooow builds

Added some rake tasks to make this easier and documented how to use them.

I've also re-enabled "destructive" mode now that we add standard arguments
to `librarian-puppet` calls, which will ensure there's no cruft left in
`vendor/modules/` and works a bit faster than calling `install --clean`.

There will be a separate pull request to reference this in gds/deployment.
But it probably needs to be merged *after* this has been promoted to
release, otherwise it will prevent us re-deploying to production in the
meantime.